# More efficient serialization of collections, `scala.Either`, and `java.time._` types