--- a/zio-json/shared/src/test/scala/zio/json/DecoderSpec.scala
+++ b/zio-json/shared/src/test/scala/zio/json/DecoderSpec.scala
@@ -4,7 +4,6 @@ import zio._
 import zio.json._
 import zio.json.ast.Json
 import zio.test.Assertion._
-import zio.test.environment.Live
 import zio.test.{ TestAspect, _ }
 
 import java.time.{ Duration, OffsetDateTime, ZonedDateTime }