--- a/zio-json/shared/src/test/scala/zio/json/EncoderSpec.scala
+++ b/zio-json/shared/src/test/scala/zio/json/EncoderSpec.scala
@@ -3,7 +3,6 @@ package testzio.json
 import zio.json._
 import zio.json.ast.Json
 import zio.test.Assertion._
-import zio.test.TestAspect._
 import zio.test._
 import zio.{ Chunk, NonEmptyChunk }