Results for OpenJDK 18 on my notebook:
## Before
```
[info] Benchmark                              (size)   Mode  Cnt         Score         Error  Units
[info] ADTWriting.zioJson                        N/A  thrpt    5    751137.135 ±   60844.184  ops/s
[info] AnyValsWriting.zioJson                    N/A  thrpt    5   1992064.842 ±  230755.542  ops/s
[info] ArrayOfBigDecimalsWriting.zioJson         128  thrpt    5     29949.979 ±     194.004  ops/s
[info] ArrayOfBooleansWriting.zioJson            128  thrpt    5    382517.605 ±    1383.240  ops/s
[info] ArrayOfBytesWriting.zioJson               128  thrpt    5    327511.459 ±   12887.265  ops/s
[info] ArrayOfCharsWriting.zioJson               128  thrpt    5    320712.508 ±   16738.339  ops/s
[info] ArrayOfDoublesWriting.zioJson             128  thrpt    5     63960.547 ±     813.024  ops/s
[info] ArrayOfDurationsWriting.zioJson           128  thrpt    5     60878.148 ±    4007.515  ops/s
[info] ArrayOfEnumADTsWriting.zioJson            128  thrpt    5    315286.974 ±   19576.020  ops/s
[info] ArrayOfEnumsWriting.zioJson               128  thrpt    5    345687.821 ±    4894.588  ops/s
[info] ArrayOfFloatsWriting.zioJson              128  thrpt    5     90803.752 ±    3823.350  ops/s
[info] ArrayOfInstantsWriting.zioJson            128  thrpt    5     55852.159 ±    4080.762  ops/s
[info] ArrayOfIntsWriting.zioJson                128  thrpt    5    232309.227 ±   13185.652  ops/s
[info] ArrayOfLocalDateTimesWriting.zioJson      128  thrpt    5     58191.175 ±    3890.360  ops/s
[info] ArrayOfLocalDatesWriting.zioJson          128  thrpt    5    116426.580 ±    7886.043  ops/s
[info] ArrayOfLocalTimesWriting.zioJson          128  thrpt    5     86121.884 ±    3637.461  ops/s
[info] ArrayOfLongsWriting.zioJson               128  thrpt    5    164070.273 ±    8967.729  ops/s
[info] ArrayOfMonthDaysWriting.zioJson           128  thrpt    5    148273.712 ±     185.215  ops/s
[info] ArrayOfOffsetDateTimesWriting.zioJson     128  thrpt    5     45906.834 ±    3632.603  ops/s
[info] ArrayOfOffsetTimesWriting.zioJson         128  thrpt    5     62093.539 ±    5495.363  ops/s
[info] ArrayOfPeriodsWriting.zioJson             128  thrpt    5     78696.749 ±    3833.590  ops/s
[info] ArrayOfShortsWriting.zioJson              128  thrpt    5    281872.036 ±   42555.978  ops/s
[info] ArrayOfUUIDsWriting.zioJson               128  thrpt    5    136503.452 ±    3772.384  ops/s
[info] ArrayOfYearMonthsWriting.zioJson          128  thrpt    5    131676.719 ±   15856.098  ops/s
[info] ArrayOfYearsWriting.zioJson               128  thrpt    5    162695.253 ±   23395.225  ops/s
[info] ArrayOfZoneIdsWriting.zioJson             128  thrpt    5    253575.942 ±   10639.567  ops/s
[info] ArrayOfZoneOffsetsWriting.zioJson         128  thrpt    5    137081.750 ±   18349.846  ops/s
[info] ArrayOfZonedDateTimesWriting.zioJson      128  thrpt    5     40339.126 ±    1209.882  ops/s
[info] Base64Writing.zioJson                     128  thrpt    5   4132074.769 ±  593890.913  ops/s
[info] GeoJSONWriting.zioJson                    N/A  thrpt    5      6187.090 ±     746.128  ops/s
[info] GitHubActionsAPIWriting.zioJson           N/A  thrpt    5    200599.336 ±     804.037  ops/s
[info] GoogleMapsAPIWriting.zioJson              N/A  thrpt    5     16176.657 ±     787.875  ops/s
[info] IntWriting.zioJson                        N/A  thrpt    5  17000257.761 ± 2535180.391  ops/s
[info] ListOfBooleansWriting.zioJson             128  thrpt    5    373433.221 ±   30617.585  ops/s
[info] MapOfIntsToBooleansWriting.zioJson        128  thrpt    5    105822.927 ±    1491.095  ops/s
[info] NestedStructsWriting.zioJson              128  thrpt    5    270221.312 ±    7824.922  ops/s
[info] PrimitivesWriting.zioJson                 N/A  thrpt    5   2352687.116 ±   86622.281  ops/s
[info] SetOfIntsWriting.zioJson                  128  thrpt    5    264131.892 ±   29876.933  ops/s
[info] StringOfAsciiCharsWriting.zioJson         128  thrpt    5   1999953.566 ±  129458.803  ops/s
[info] StringOfNonAsciiCharsWriting.zioJson      128  thrpt    5    910241.118 ±  143804.693  ops/s
[info] VectorOfBooleansWriting.zioJson           128  thrpt    5    413664.436 ±    2254.215  ops/s
```
## After
```
[info] Benchmark                              (size)   Mode  Cnt         Score         Error  Units
[info] ADTWriting.zioJson                        N/A  thrpt    5    765205.782 ±   45223.173  ops/s
[info] AnyValsWriting.zioJson                    N/A  thrpt    5   2003174.417 ±  176973.603  ops/s
[info] ArrayOfBigDecimalsWriting.zioJson         128  thrpt    5     30297.028 ±    1656.736  ops/s
[info] ArrayOfBooleansWriting.zioJson            128  thrpt    5    458533.793 ±   28382.429  ops/s
[info] ArrayOfBytesWriting.zioJson               128  thrpt    5    357977.852 ±   20938.364  ops/s
[info] ArrayOfCharsWriting.zioJson               128  thrpt    5    456212.528 ±   27073.913  ops/s
[info] ArrayOfDoublesWriting.zioJson             128  thrpt    5     65973.075 ±     630.854  ops/s
[info] ArrayOfDurationsWriting.zioJson           128  thrpt    5     65292.186 ±    2181.715  ops/s
[info] ArrayOfEnumADTsWriting.zioJson            128  thrpt    5    312215.415 ±   27536.009  ops/s
[info] ArrayOfEnumsWriting.zioJson               128  thrpt    5    361466.648 ±    5993.181  ops/s
[info] ArrayOfFloatsWriting.zioJson              128  thrpt    5     95933.208 ±    1062.174  ops/s
[info] ArrayOfInstantsWriting.zioJson            128  thrpt    5     92459.782 ±     250.623  ops/s
[info] ArrayOfIntsWriting.zioJson                128  thrpt    5    238332.597 ±   16080.859  ops/s
[info] ArrayOfLocalDateTimesWriting.zioJson      128  thrpt    5     65826.134 ±    3189.740  ops/s
[info] ArrayOfLocalDatesWriting.zioJson          128  thrpt    5    189121.924 ±   13587.510  ops/s
[info] ArrayOfLocalTimesWriting.zioJson          128  thrpt    5    144061.944 ±    2350.764  ops/s
[info] ArrayOfLongsWriting.zioJson               128  thrpt    5    175362.727 ±     211.747  ops/s
[info] ArrayOfMonthDaysWriting.zioJson           128  thrpt    5    245150.019 ±   11447.880  ops/s
[info] ArrayOfOffsetDateTimesWriting.zioJson     128  thrpt    5     52415.651 ±    4449.709  ops/s
[info] ArrayOfOffsetTimesWriting.zioJson         128  thrpt    5    111603.882 ±    7455.122  ops/s
[info] ArrayOfPeriodsWriting.zioJson             128  thrpt    5     80609.331 ±    2497.186  ops/s
[info] ArrayOfShortsWriting.zioJson              128  thrpt    5    295437.180 ±   24011.888  ops/s
[info] ArrayOfUUIDsWriting.zioJson               128  thrpt    5    140624.591 ±     808.847  ops/s
[info] ArrayOfYearMonthsWriting.zioJson          128  thrpt    5    227914.680 ±    1915.558  ops/s
[info] ArrayOfYearsWriting.zioJson               128  thrpt    5    285197.228 ±     473.101  ops/s
[info] ArrayOfZoneIdsWriting.zioJson             128  thrpt    5    267295.825 ±   10305.064  ops/s
[info] ArrayOfZoneOffsetsWriting.zioJson         128  thrpt    5    188215.851 ±    5746.485  ops/s
[info] ArrayOfZonedDateTimesWriting.zioJson      128  thrpt    5     55637.488 ±    2871.643  ops/s
[info] Base64Writing.zioJson                     128  thrpt    5   4053856.014 ±   13990.812  ops/s
[info] GeoJSONWriting.zioJson                    N/A  thrpt    5      6307.678 ±     534.233  ops/s
[info] GitHubActionsAPIWriting.zioJson           N/A  thrpt    5    201647.170 ±    1165.914  ops/s
[info] GoogleMapsAPIWriting.zioJson              N/A  thrpt    5     16444.251 ±     962.177  ops/s
[info] IntWriting.zioJson                        N/A  thrpt    5  17553304.162 ± 1666812.485  ops/s
[info] ListOfBooleansWriting.zioJson             128  thrpt    5    403899.956 ±   25586.063  ops/s
[info] MapOfIntsToBooleansWriting.zioJson        128  thrpt    5    111123.998 ±     825.866  ops/s
[info] NestedStructsWriting.zioJson              128  thrpt    5    263101.546 ±   27989.424  ops/s
[info] PrimitivesWriting.zioJson                 N/A  thrpt    5   2337839.699 ±  253900.390  ops/s
[info] SetOfIntsWriting.zioJson                  128  thrpt    5    257074.244 ±    4841.420  ops/s
[info] StringOfAsciiCharsWriting.zioJson         128  thrpt    5   1867367.087 ±  131764.617  ops/s
[info] StringOfNonAsciiCharsWriting.zioJson      128  thrpt    5    957493.457 ±   69881.943  ops/s
[info] VectorOfBooleansWriting.zioJson           128  thrpt    5    408153.991 ±   45155.071  ops/s
```